backend/requirements.txt
# Core Django Framework
Django==5.0.1

# Database
psycopg2-binary==2.9.11

# Environment Variables Management
python-decouple==3.8

# Additional utilities (will add more as we progress)
pytz==2024.1

# Django REST Framework (API)
djangorestframework==3.14.0

# CORS (Cross-Origin Resource Sharing) - Required for React frontend
django-cors-headers==4.3.1

# JWT Authentication for API
djangorestframework-simplejwt==5.3.0

# Celery for async tasks (PDF processing)
celery==5.3.4
redis==5.0.1  # Message broker for Celery

# File handling
Pillow==10.1.0  # Image processing (if needed for file previews)

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0  # Coverage reports

# PDF processing (for extracting text from PDFs)
PyPDF2==3.0.1
python-docx==1.1.0  # For DOCX file processing

# OpenAI API for AI-powered contract analysis
openai>=1.0.0

